Moving navigation into separate module
