+++ app/src/main/java/com/jraska/github/client/ChromeCustomTabsModule.kt
@@ -1,0 +1,16 @@
+package com.jraska.github.client
+
+import com.jraska.github.client.core.android.TopActivityProvider
+import com.jraska.github.client.rx.AppSchedulers
+import dagger.Module
+import dagger.Provides
+
+@Module
+object ChromeCustomTabsModule {
+
+  @JvmStatic
+  @Provides
+  fun webLinkLauncher(provider: TopActivityProvider, appSchedulers: AppSchedulers): WebLinkLauncher {
+    return ChromeCustomTabsLauncher(provider, appSchedulers)
+  }
+}
