+++ feature/users/build.gradle
@@ -18,6 +18,7 @@ android {
 dependencies {
   api project(':core')
   api project(':core-android')
+  api project(':navigation')
 
   kapt 'com.google.dagger:dagger-compiler:2.24'
   api 'com.google.dagger:dagger:2.24'
