+++ lib/navigation-deeplink/build.gradle
@@ -15,19 +15,13 @@ android {
 }
 
 dependencies {
+  api project(':navigation')
   api project(':core')
-
-  api 'androidx.appcompat:appcompat:1.1.0'
-  api 'androidx.annotation:annotation:1.1.0'
-  api 'androidx.lifecycle:lifecycle-extensions:2.1.0'
-  api 'androidx.lifecycle:lifecycle-reactivestreams-ktx:2.1.0'
+  api project(':core-android')
 
   kapt 'com.google.dagger:dagger-compiler:2.24'
-  implementation 'com.google.dagger:dagger:2.24'
-
-  implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
-
-  api 'com.jakewharton.timber:timber:4.7.1'
+  api 'com.google.dagger:dagger:2.24'
 
   testImplementation 'junit:junit:4.12'
+  testImplementation 'org.mockito:mockito-core:3.0.0'
 }
