+++ lib/navigation-deeplink/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.jraska.github.client.navigation.deeplink" />
